+++ app/scenes/Collection.js
@@ -27,10 +27,10 @@ import Flex from "components/Flex";
 import Heading from "components/Heading";
 import HelpText from "components/HelpText";
 import InputSearchPage from "components/InputSearchPage";
+import PlaceholderList from "components/List/Placeholder";
 import LoadingIndicator from "components/LoadingIndicator";
 import Modal from "components/Modal";
 import PaginatedDocumentList from "components/PaginatedDocumentList";
-import PlaceholderList from "components/PlaceholderList";
 import PlaceholderText from "components/PlaceholderText";
 import Scene from "components/Scene";
 import Subheading from "components/Subheading";
